moved links to navBar
- moved links for home, about and contact page to navBar for now
